feat: Add Scout assignment to Troop functionality
Backend changes:
- Add getScoutsByTroop endpoint (GET /users/troop/{troopId}/scouts)
- Add getUnassignedScouts endpoint (GET /users/scouts/unassigned)
- Add assignScoutToTroop endpoint (PATCH /users/{userId}/assign-troop/{troopId})
- Add removeScoutFromTroop endpoint (DELETE /users/{userId}/troop)
- Add corresponding methods to UserService and UserRepository
- Fix Integer vs Long type mismatch in CouncilService

Frontend changes:
- Add getScoutsByTroop, getUnassignedScouts API methods
- Add assignScoutToTroop, removeScoutFromTroop API methods

+++ backend/src/main/java/com/bsa/campcard/service/CouncilService.java
@@ -316,11 +316,11 @@ public class CouncilService {
         // Get troop counts and stats
         Long troopCount = troopRepository.countByCouncilId(councilId);
         BigDecimal totalSales = troopRepository.sumSalesByCouncil(councilId);
-        Long cardsSold = troopRepository.sumCardsSoldByCouncil(councilId);
+        Integer cardsSold = troopRepository.sumCardsSoldByCouncil(councilId);
 
         council.setTotalTroops(troopCount != null ? troopCount.intValue() : 0);
         council.setTotalSales(totalSales != null ? totalSales : BigDecimal.ZERO);
-        council.setCardsSold(cardsSold != null ? cardsSold.intValue() : 0);
+        council.setCardsSold(cardsSold != null ? cardsSold : 0);
 
         Council saved = councilRepository.save(council);
         log.info("Council stats updated: troops={}, sales={}, cards={}",
